requirements.txt
kafka-python==2.0.2
websocket-client==1.8.0
python-dotenv==1.0.1
requests==2.32.3
pandas==1.5.3
numpy==1.26.4
pyarrow==10.0.1
pyspark==3.3.2
delta-spark==2.3.0
streamlit==1.36.0
deltalake==0.17.4
pytest==7.3.1
plotly==5.22.0